+++ packages/api/endpoints/pdrs.js
@@ -78,9 +78,29 @@ async function del(req, res) {
 
   const pdrName = req.params.pdrName;
   const pdrS3Key = `${process.env.stackName}/pdrs/${pdrName}`;
+  const esPdrsClient = new Search(
+    {},
+    'pdr',
+    process.env.ES_INDEX
+  );
 
   let existingPdr;
   try {
+    await pdrPgModel.get(knex, { name: pdrName });
+  } catch (error) {
+    if (error instanceof RecordDoesNotExist) {
+      if (!(await esPdrsClient.exists(pdrName))) {
+        log.info('PDR does not exist in Elasticsearch');
+        return res.boom.notFound('No record found');
+      }
+      log.info('PDR does not exist in PostgreSQL, it only exists in Elasticsearch');
+    } else {
+      throw error;
+    }
+  }
+
+  try {
+    // Save DynamoDb PDR in case delete fails and we need to recreate
     existingPdr = await pdrModel.get({ pdrName });
   } catch (error) {
     // Ignore error if record does not exist in DynamoDb